Refactor + Config + Drive with heading
