+++ src/main/java/swervelib/Constants.java
@@ -2,6 +2,8 @@ package swervelib;
 
 public final class Constants {
   // Will fix some issues with abs positions when abs pos fails
-  public final static boolean ENABLE_ABS_ENCODER_POSITION = true;
+  public final static boolean ENABLE_ABS_ENCODER_POS_ERROR_CHECKS = true;
+  public final static int ABS_ENCODER_ERROR_RETRY_COUNT = 3;
 
+  public final static boolean BOOT_TO_ABS = true;
 }
