+++ src/main/java/swervelib/ModuleConfiguration.java
@@ -2,6 +2,8 @@ package swervelib;
 
 import java.util.Objects;
 
+import swervelib.config.SdsModuleConfigurations;
+
 /**
  * A swerve module configuration.
  * <p>
