+++ src/main/java/swervelib/config/Mk3ModuleConfiguration.java
@@ -1,14 +1,14 @@
-package swervelib;
+package swervelib.config;
 
 import java.util.Objects;
 
 /**
- * Additional Mk4 module configuration parameters.
+ * Additional Mk3 module configuration parameters.
  * <p>
- * The configuration parameters here are used to customize the behavior of the Mk4 swerve module.
+ * The configuration parameters here are used to customize the behavior of the Mk3 swerve module.
  * Each setting is initialized to a default that should be adequate for most use cases.
  */
-public class Mk4ModuleConfiguration {
+public class Mk3ModuleConfiguration {
     private double nominalVoltage = 12.0;
     private double driveCurrentLimit = 80.0;
     private double steerCurrentLimit = 20.0;
@@ -43,7 +43,7 @@ public class Mk4ModuleConfiguration {
             return true;
         if (o == null || getClass() != o.getClass())
             return false;
-        Mk4ModuleConfiguration that = (Mk4ModuleConfiguration) o;
+        Mk3ModuleConfiguration that = (Mk3ModuleConfiguration) o;
         return Double.compare(that.getNominalVoltage(), getNominalVoltage()) == 0
                 && Double.compare(that.getDriveCurrentLimit(), getDriveCurrentLimit()) == 0
                 && Double.compare(that.getSteerCurrentLimit(), getSteerCurrentLimit()) == 0;
@@ -56,7 +56,7 @@ public class Mk4ModuleConfiguration {
 
     @Override
     public String toString() {
-        return "Mk4ModuleConfiguration{" +
+        return "Mk3ModuleConfiguration{" +
                 "nominalVoltage=" + nominalVoltage +
                 ", driveCurrentLimit=" + driveCurrentLimit +
                 ", steerCurrentLimit=" + steerCurrentLimit +
