+++ src/main/java/swervelib/config/Mk4SwerveModuleHelper.java
@@ -1,15 +1,20 @@
-package swervelib;
+package swervelib.config;
 
+import swervelib.DriveControllerFactory;
+import swervelib.ModuleConfiguration;
+import swervelib.SteerControllerFactory;
+import swervelib.SwerveModule;
+import swervelib.SwerveModuleFactory;
 import swervelib.ctre.*;
 import swervelib.rev.*;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardLayout;
 
-public final class Mk3SwerveModuleHelper {
-        private Mk3SwerveModuleHelper() {
+public final class Mk4SwerveModuleHelper {
+        private Mk4SwerveModuleHelper() {
         }
 
         private static DriveControllerFactory<?, Integer> getFalcon500DriveFactory(
-                        Mk3ModuleConfiguration configuration) {
+                        Mk4ModuleConfiguration configuration) {
                 return new Falcon500DriveControllerFactoryBuilder()
                                 .withVoltageCompensation(configuration.getNominalVoltage())
                                 .withCurrentLimit(configuration.getDriveCurrentLimit())
@@ -17,7 +22,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         private static SteerControllerFactory<?, Falcon500SteerConfiguration<CanCoderAbsoluteConfiguration>> getFalcon500SteerFactory(
-                        Mk3ModuleConfiguration configuration) {
+                        Mk4ModuleConfiguration configuration) {
                 return new Falcon500SteerControllerFactoryBuilder()
                                 .withVoltageCompensation(configuration.getNominalVoltage())
                                 .withPidConstants(0.2, 0.0, 0.1)
@@ -27,7 +32,7 @@ public final class Mk3SwerveModuleHelper {
                                                 .build());
         }
 
-        private static DriveControllerFactory<?, Integer> getNeoDriveFactory(Mk3ModuleConfiguration configuration) {
+        private static DriveControllerFactory<?, Integer> getNeoDriveFactory(Mk4ModuleConfiguration configuration) {
                 return new NeoDriveControllerFactoryBuilder()
                                 .withVoltageCompensation(configuration.getNominalVoltage())
                                 .withCurrentLimit(configuration.getDriveCurrentLimit())
@@ -35,7 +40,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         private static SteerControllerFactory<?, NeoSteerConfiguration<CanCoderAbsoluteConfiguration>> getNeoSteerFactory(
-                        Mk3ModuleConfiguration configuration) {
+                        Mk4ModuleConfiguration configuration) {
                 return new NeoSteerControllerFactoryBuilder()
                                 .withVoltageCompensation(configuration.getNominalVoltage())
                                 .withPidConstants(1.0, 0.0, 0.1)
@@ -46,7 +51,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
+         * Creates a Mk4 swerve module that uses Falcon 500s for driving and steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -60,7 +65,7 @@ public final class Mk3SwerveModuleHelper {
          */
         public static SwerveModule createFalcon500(
                         ShuffleboardLayout container,
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -79,7 +84,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
+         * Creates a Mk4 swerve module that uses Falcon 500s for driving and steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -97,12 +102,12 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createFalcon500(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort,
+                return createFalcon500(container, new Mk4ModuleConfiguration(), gearRatio, driveMotorPort,
                                 steerMotorPort, steerEncoderPort, steerOffset);
         }
 
         /**
-         * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
+         * Creates a Mk4 swerve module that uses Falcon 500s for driving and steering.
          *
          * @param configuration    Module configuration parameters to use.
          * @param gearRatio        The gearing configuration the module is in.
@@ -113,7 +118,7 @@ public final class Mk3SwerveModuleHelper {
          * @return The configured swerve module.
          */
         public static SwerveModule createFalcon500(
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -131,7 +136,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
+         * Creates a Mk4 swerve module that uses Falcon 500s for driving and steering.
          *
          * @param gearRatio        The gearing configuration the module is in.
          * @param driveMotorPort   The CAN ID of the drive Falcon 500.
@@ -146,12 +151,12 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createFalcon500(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
+                return createFalcon500(new Mk4ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
                                 steerEncoderPort, steerOffset);
         }
 
         /**
-         * Creates a Mk3 swerve module that uses NEOs for driving and steering.
+         * Creates a Mk4 swerve module that uses NEOs for driving and steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -165,7 +170,7 @@ public final class Mk3SwerveModuleHelper {
          */
         public static SwerveModule createNeo(
                         ShuffleboardLayout container,
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -184,7 +189,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses NEOs for driving and steering.
+         * Creates a Mk4 swerve module that uses NEOs for driving and steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -202,12 +207,12 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createNeo(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
+                return createNeo(container, new Mk4ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
                                 steerEncoderPort, steerOffset);
         }
 
         /**
-         * Creates a Mk3 swerve module that uses NEOs for driving and steering.
+         * Creates a Mk4 swerve module that uses NEOs for driving and steering.
          *
          * @param configuration    Module configuration parameters to use.
          * @param gearRatio        The gearing configuration the module is in.
@@ -218,7 +223,7 @@ public final class Mk3SwerveModuleHelper {
          * @return The configured swerve module.
          */
         public static SwerveModule createNeo(
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -236,7 +241,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses NEOs for driving and steering.
+         * Creates a Mk4 swerve module that uses NEOs for driving and steering.
          *
          * @param gearRatio        The gearing configuration the module is in.
          * @param driveMotorPort   The CAN ID of the drive NEO.
@@ -251,12 +256,12 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createNeo(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
+                return createNeo(new Mk4ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
                                 steerEncoderPort, steerOffset);
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
+         * Creates a Mk4 swerve module that uses a Falcon 500 for driving and a NEO for steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -270,7 +275,7 @@ public final class Mk3SwerveModuleHelper {
          */
         public static SwerveModule createFalcon500Neo(
                         ShuffleboardLayout container,
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -289,7 +294,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
+         * Creates a Mk4 swerve module that uses a Falcon 500 for driving and a NEO for steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -307,12 +312,12 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createFalcon500Neo(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort,
+                return createFalcon500Neo(container, new Mk4ModuleConfiguration(), gearRatio, driveMotorPort,
                                 steerMotorPort, steerEncoderPort, steerOffset);
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
+         * Creates a Mk4 swerve module that uses a Falcon 500 for driving and a NEO for steering.
          *
          * @param configuration    Module configuration parameters to use.
          * @param gearRatio        The gearing configuration the module is in.
@@ -323,7 +328,7 @@ public final class Mk3SwerveModuleHelper {
          * @return The configured swerve module.
          */
         public static SwerveModule createFalcon500Neo(
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -341,7 +346,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
+         * Creates a Mk4 swerve module that uses a Falcon 500 for driving and a NEO for steering.
          *
          * @param gearRatio        The gearing configuration the module is in.
          * @param driveMotorPort   The CAN ID of the drive Falcon 500.
@@ -356,12 +361,12 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createFalcon500Neo(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
+                return createFalcon500Neo(new Mk4ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
                                 steerEncoderPort, steerOffset);
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
+         * Creates a Mk4 swerve module that uses a NEO for driving and a Falcon 500 for steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -375,7 +380,7 @@ public final class Mk3SwerveModuleHelper {
          */
         public static SwerveModule createNeoFalcon500(
                         ShuffleboardLayout container,
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -394,7 +399,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
+         * Creates a Mk4 swerve module that uses a NEO for driving and a Falcon 500 for steering.
          * Module information is displayed in the specified ShuffleBoard container.
          *
          * @param container        The container to display module information in.
@@ -412,12 +417,12 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createNeoFalcon500(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort,
+                return createNeoFalcon500(container, new Mk4ModuleConfiguration(), gearRatio, driveMotorPort,
                                 steerMotorPort, steerEncoderPort, steerOffset);
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
+         * Creates a Mk4 swerve module that uses a NEO for driving and a Falcon 500 for steering.
          *
          * @param configuration    Module configuration parameters to use.
          * @param gearRatio        The gearing configuration the module is in.
@@ -428,7 +433,7 @@ public final class Mk3SwerveModuleHelper {
          * @return The configured swerve module.
          */
         public static SwerveModule createNeoFalcon500(
-                        Mk3ModuleConfiguration configuration,
+                        Mk4ModuleConfiguration configuration,
                         GearRatio gearRatio,
                         int driveMotorPort,
                         int steerMotorPort,
@@ -446,7 +451,7 @@ public final class Mk3SwerveModuleHelper {
         }
 
         /**
-         * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
+         * Creates a Mk4 swerve module that uses a NEO for driving and a Falcon 500 for steering.
          *
          * @param gearRatio        The gearing configuration the module is in.
          * @param driveMotorPort   The CAN ID of the drive NEO.
@@ -461,19 +466,15 @@ public final class Mk3SwerveModuleHelper {
                         int steerMotorPort,
                         int steerEncoderPort,
                         double steerOffset) {
-                return createNeoFalcon500(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
+                return createNeoFalcon500(new Mk4ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort,
                                 steerEncoderPort, steerOffset);
         }
 
         public enum GearRatio {
-                /**
-                 * Mk3 swerve in the standard gear configuration.
-                 */
-                STANDARD(SdsModuleConfigurations.MK3_STANDARD),
-                /**
-                 * Mk3 swerve in the fast gear configuration.
-                 */
-                FAST(SdsModuleConfigurations.MK3_FAST);
+                L1(SdsModuleConfigurations.MK4_L1),
+                L2(SdsModuleConfigurations.MK4_L2),
+                L3(SdsModuleConfigurations.MK4_L3),
+                L4(SdsModuleConfigurations.MK4_L4);
 
                 private final ModuleConfiguration configuration;
