+++ src/main/java/swervelib/ctre/CanCoderAbsoluteConfiguration.java
@@ -2,13 +2,16 @@ package swervelib.ctre;
 
 import com.ctre.phoenix.sensors.SensorInitializationStrategy;
 
+import swervelib.Constants;
+
 public class CanCoderAbsoluteConfiguration {
     private final int id;
     private final double offset;
     private final SensorInitializationStrategy initStrategy;
 
     public CanCoderAbsoluteConfiguration(int id, double offset) {
-        this(id, offset, SensorInitializationStrategy.BootToAbsolutePosition);
+        this(id, offset, Constants.BOOT_TO_ABS ? SensorInitializationStrategy.BootToAbsolutePosition
+                : SensorInitializationStrategy.BootToZero);
     }
 
     public CanCoderAbsoluteConfiguration(int id, double offset, SensorInitializationStrategy initStrategy) {
